+++ CHANGELOG.md
@@ -11,7 +11,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 ## [Unreleased]
 ## [0.2.0-alpha1] - 2024-10-19
 
-- switch to half-precision training and nightly Rust
+- allow half-precision training
+- switch to nightly Rust
 
 ## [0.1.0] - 2024-10-19