+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 ## [0.2.0-alpha1] - 2024-10-19
 
 - allow half-precision training
+- dynamicly get binary name when generating completion
 
 ## [0.1.0] - 2024-10-19